+++ src/test/java/com/maxmind/minfraud/response/DispositionTest.java
@@ -3,8 +3,6 @@ package com.maxmind.minfraud.response;
 import com.fasterxml.jackson.jr.ob.JSON;
 import org.junit.Test;
 
-import java.util.UUID;
-
 import static org.junit.Assert.assertEquals;
 
 public class DispositionTest extends AbstractOutputTest {